Stop a refused template edit from looking like a saved one
Every template save was being rejected by Postgres with 22003 and nothing said
so. The column shipped as `sort_order integer` while the client stamps a new
row with `Date.now()` — 1787914309491, three orders of magnitude past int4 —
so the insert never landed. Hiding a template worked, because that path passes
0, which is exactly why the bug looked like "editing specifically is broken".

Two fixes, because the overflow was only half of it. The column is bigint now,
like every other sort_order in the schema, which is what it should have been.
And the write no longer swallows its own failure: the optimistic row is rolled
back, the error is rethrown, and the form stays open with the text still in the
fields rather than closing as though the edit had been kept. A save that cannot
succeed should never look identical to one that did — that is what turned a
400 into "it doesn't save the edited template" with nothing to go on.

The merge behind the picker gets tests. It reconciles constants that ship with
the app against rows that exist only where one has been changed, and every way
that can go wrong is silent: a hidden built-in reappearing, a custom one
sorting into the middle, reset restoring the wrong text.

Earlier notes gets around three more rows before it starts scrolling.

+++ supabase/schema.sql
@@ -163,7 +163,9 @@ create table if not exists public.clarity_note_templates (
   -- A built-in the reader has put away. The row exists so the choice syncs;
   -- deleting the row is how a built-in goes back to its shipped shape.
   hidden      boolean not null default false,
-  sort_order  integer not null default 0,
+  -- bigint, like every other sort_order here: the client stamps these with
+  -- Date.now(), which overflows int4 and made every save a silent 400.
+  sort_order  bigint not null default 0,
   created_at  timestamptz not null default now(),
   updated_at  timestamptz not null default now()
 );
